+++ src/pages/Home/index.tsx
@@ -7,7 +7,6 @@ import Img1 from '@assets/images/football1.jpg';
 import Img2 from '@assets/images/football2.jpg';
 import Img3 from '@assets/images/football3.jpg';
 import CommonLayout from '@components/commonLayout';
-import { testApi } from '@service/matchApi';
 import '@assets/home.scss';
 
 import MatchSelector from './matchSelector';
@@ -46,7 +45,6 @@ const Home = () => {
       <MatchContainer>
         <MatchSelector />
       </MatchContainer>
-      <button onClick={() => testApi()}>테스트</button>
     </CommonLayout>
   );
 };
