--- src/service/authHeader.ts
@@ -1,6 +1,0 @@
-const authHeader = () => {
-  const token = localStorage.getItem('token') ?? '';
-  return { headers: { Authorization: `Bearer${token}` } };
-};
-
-export default authHeader;
