+++ src/service/config.ts
@@ -1,2 +1,2 @@
-// export const apiUrl = 'https://football-matching.herokuapp.com';
-export const apiUrl = 'http://localhost:8080';
+export const apiUrl = 'https://football-matching.herokuapp.com';
+// export const apiUrl = 'http://localhost:8080';
