+++ src/service/matchApi.ts
@@ -1,6 +1,7 @@
 import axios, { AxiosError, AxiosRequestConfig, AxiosResponse } from 'axios';
 
-import authHeader from './authHeader';
+import type { MatchInfo, MatchKeys } from '@custype/matchTypes';
+
 import { apiUrl } from './config';
 
 const matchAxios = axios.create({ baseURL: `${apiUrl}/matches` });
@@ -12,37 +13,21 @@ matchAxios.interceptors.request.use((config: AxiosRequestConfig) => {
   };
   return config;
 });
+matchAxios.interceptors.response.use(
+  (response) => response,
+  (error) => {
+    return Promise.reject(error);
+  },
+);
 
-interface matchInfo {
-  stadiumId: number;
-  matchNum: number;
-  gender: string;
-  content: string;
-  startAt: string;
-}
-
-interface MatchKeys {
-  matchDay?: string;
-  gender?: string;
-  status?: string;
-  personnel?: number;
-  stadiumName?: string;
-}
-
-// export const createMatch = (matchInfo: matchInfo) => {
-//   return matchAxios.post(`${apiUrl}/matches`, matchInfo, authHeader());
-// };
-
-export const createMatch = async (matchInfo: matchInfo) => {
-  return await wrapper(matchAxios.post('', matchInfo));
+export const createMatch = async (matchInfo: MatchInfo) => {
+  return await matchAxios.post('', matchInfo);
 };
 
 export const getMatches = async (queryString: MatchKeys): Promise<any[]> => {
-  const res = await wrapper(
-    matchAxios.get(``, {
-      params: queryString,
-    }),
-  );
+  const res = await matchAxios.get(``, {
+    params: queryString,
+  });
   if (res.data) {
     return res.data.content;
   } else {
@@ -51,27 +36,21 @@ export const getMatches = async (queryString: MatchKeys): Promise<any[]> => {
 };
 
 export const getMatch = async (matchId: number) => {
-  const res = await wrapper(matchAxios.get(`${matchId}`));
+  const res = await matchAxios.get(`${matchId}`);
   if (res) return res.data;
   else return null;
 };
 
-export const testApi = () => {
-  return matchAxios.get(`/test`);
-};
-
-const wrapper = async (
-  myFunc: Promise<AxiosResponse<any, any>>,
-): Promise<any> => {
-  try {
-    return await myFunc;
-  } catch (error) {
-    if (error instanceof AxiosError) {
-      console.error(error.message);
-      return error.message;
-    } else {
-      console.error(error);
-      return error;
-    }
-  }
-};
+// const wrapper = async (
+//   myFunc: Promise<AxiosResponse<any, any>>,
+// ): Promise<any> => {
+//   try {
+//     return await myFunc;
+//   } catch (error) {
+//     if (error instanceof AxiosError) {
+//       return await Promise.reject(error);
+//     } else {
+//       return await Promise.reject(error);
+//     }
+//   }
+// };
