+++ src/service/stadiumApi.ts
@@ -1,6 +1,5 @@
 import axios, { AxiosRequestConfig } from 'axios';
 
-import authHeader from './authHeader';
 import { apiUrl } from './config';
 
 const stadiumAxios = axios.create({ baseURL: `${apiUrl}/` });
@@ -13,10 +12,6 @@ stadiumAxios.interceptors.request.use((config: AxiosRequestConfig) => {
   return config;
 });
 
-// Question
-// const token = authHeader()을 함수 실행 전 매번 붙여줘야 한다.
-// 매번 안붙이고 선행으로 authHeader값을 선언하지 않고 하는 방법
-
 export const createStadium = (stadiumInfo: {
   name: string;
   content: string;
@@ -25,7 +20,7 @@ export const createStadium = (stadiumInfo: {
   address: string;
   imageUrl: string;
 }) => {
-  return axios.post(`${apiUrl}/stadiums`, stadiumInfo, authHeader());
+  return axios.post(`${apiUrl}/stadiums`, stadiumInfo);
 };
 
 export const getStadiumList = (address: string) => {
