+++ src/service/userApi.ts
@@ -1,6 +1,5 @@
 import axios from 'axios';
 
-import authHeader from './authHeader';
 import { apiUrl } from './config';
 
 export interface UserInfo {
